requirements.txt
# --- Framework Principal ---
django==5.0.6

# --- Inteligência Artificial e Análise de Dados ---
pandasai==2.3.2
pandas==1.5.3
numpy==1.26.4
google-generativeai==0.5.4
SQLAlchemy==2.0.29 # Dependência do PandasAI para conexão com banco de dados
tabulate==0.9.0    # Dependência do Pandas para formatação de tabelas

# --- Ferramentas e Utilitários do Django ---
django-cors-headers==4.3.1 # Para permitir a comunicação entre front-end e back-end

# --- Outras dependências ---
PyYAML==6.0.1 # Dependência do PandasAI